--- app/src/main/java/cp17304_n3/fpoly/du_an_1/DAO/boss.java
@@ -1,4 +1,0 @@
-package cp17304_n3.fpoly.du_an_1.DAO;
-
-public class boss {
-}
