--- app/src/main/java/cp17304_n3/fpoly/du_an_1/DTO/Boss.java
@@ -1,8 +1,0 @@
-package cp17304_n3.fpoly.du_an_1.DTO;
-
-public class Boss {
-int idBoss;
-String hoten;
-String user;
-String pass;
-}
