+++ app/src/main/java/cp17304_n3/fpoly/du_an_1/Datbase/DAO/boss.java
@@ -1,0 +1,4 @@
+package cp17304_n3.fpoly.du_an_1.Datbase.DAO;
+
+public class boss {
+}
