+++ app/src/main/java/cp17304_n3/fpoly/du_an_1/Fragment/ThongKe.java
@@ -23,15 +23,6 @@ public class ThongKe extends Fragment {
         // Required empty public constructor
     }
 
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment ThongKe.
-     */
-    // TODO: Rename and change types and number of parameters
     public static ThongKe newInstance() {
         ThongKe fragment = new ThongKe();
