+++ app/src/main/java/cp17304_n3/fpoly/du_an_1/Fragment/danhSachDonHang.java
@@ -17,16 +17,6 @@ public class danhSachDonHang extends Fragment {
     public danhSachDonHang() {
         // Required empty public constructor
     }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment danhSachDonHang.
-     */
-    // TODO: Rename and change types and number of parameters
     public static danhSachDonHang newInstance() {
         danhSachDonHang fragment = new danhSachDonHang();
